+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Version 0.0.320 (February 24, 2026)
+- Improve table lookup
+- Improve OAuth MCP integration
+
 ## Version 0.0.320 (February 22, 2026)
 - Support deployment in airgapped systems
 - Improve PowerBI integration